+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 -   Use `xiblepm nodepack init` to create a default node structure. This is useful for starting the development of a new node within a nodepack. ([#81](https://github.com/SpectrumBroad/xible/issues/81))
 
+-   Unsupported browsers can now find exactly which features the browser is missing in a 'details' window.
+
 ### Fixed
 -   When publishing a nodepack (which already exists), the package.json `name` attribute is now stripped of preceeding 'xible-np-' or 'xible-nodepack-' strings before matching it against the registry to see if the publisher is the same. Prior to this, the registry would return a http 403 which would show up as 'Unsuccessful statuscode returned' in the cli.